// frontend/src/lib/api/model/createUserBody.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * web-app-template API
 * REST API for web-app-template.
 * OpenAPI spec version: 0.0.0
 */

export type CreateUserBody = {
  /**
   * User email address
   * @minLength 5
   * @maxLength 320
   */
  email: string;
  /**
   * Country of residence
   * @minLength 4
   * @maxLength 50
   */
  country: string;
  /**
   * Postal code
   * @minLength 4
   * @maxLength 10
   */
  postalCode: string;
  /**
   * Street
   * @minLength 5
   * @maxLength 50
   */
  street: string;
};